Fixed a transparency issue (TextMate)
When the current line was highlighted, the right margin indicator was activated and the right margin was highlighted, there was a transparency issue. The current line highlight overlapped the right margin highlight.
I added a 50% transparency and changed the colors of the current line highlight, so they match the previous color when being displayed on top of the background color.

Example image: http://goo.gl/bzlnD